+++ respy/simulate.py
@@ -12,7 +12,7 @@ from respy.shared import create_base_draws
 from respy.shared import get_conditional_probabilities
 from respy.shared import transform_disturbances
 from respy.solve import solve_with_backward_induction
-from respy.solve import StateSpace
+from respy.state_space import StateSpace
 
 
 def simulate(params, options):
