+++ respy/solve.py
@@ -1,15 +1,13 @@
 import numpy as np
-import pandas as pd
 from numba import guvectorize
 from numba import njit
 
-from respy.config import BASE_COVARIATES
 from respy.config import HUGE_FLOAT
 from respy.pre_processing.model_processing import process_options
 from respy.pre_processing.model_processing import process_params
 from respy.shared import _aggregate_keane_wolpin_utility
-from respy.shared import create_base_draws
 from respy.shared import transform_disturbances
+from respy.state_space import StateSpace
 
 
 def solve(params, options):
@@ -34,241 +32,6 @@ def solve(params, options):
     )
 
     return state_space
-
-
-@njit
-def create_state_space(num_periods, num_types, edu_starts, edu_max):
-    """Create the state space.
-
-    The state space consists of all admissible combinations of the following elements:
-    period, experience in OCCUPATION A, experience in OCCUPATION B, years of schooling,
-    the lagged choice and the type of the agent.
-
-    The major problem which is solved here is that the natural representation of the
-    state space is a graph where each parent node is connected with its children's
-    nodes. This structure makes it easy to traverse the graph. In contrast to that, for
-    many subsequent calculations, e.g. generating the covariates for each state, a
-    tabular format has better performance, but looses the information on the
-    connections.
-
-    The current implementation of ``states`` and ``indexer`` allows to have both
-    advantages at the cost of an additional object. ``states`` stores the information on
-    states in a tabular format. ``indexer`` is a matrix where each characteristic of the
-    state space represents one dimension. The values of the matrix are the indices of
-    states in ``states``. Traversing the state space is as easy as incrementing the
-    right indices of ``indexer`` by 1 and use the resulting index in ``states``.
-
-    Parameters
-    ----------
-    num_periods : int
-        Number of periods in the state space.
-    num_types : int
-        Number of types of agents.
-    edu_starts : List[int]
-        Contains levels of initial education.
-    edu_max : int
-        Maximum level of education which can be obtained by an agent.
-
-    Returns
-    -------
-    states : np.ndarray
-        Array with shape (num_states, 6) containing period, experience in OCCUPATION A,
-        experience in OCCUPATION B, years of schooling, the lagged choice and the type
-        of the agent.
-    indexer : np.ndarray
-        A matrix where each dimension represents a characteristic of the state space.
-        Switching from one state is possible via incrementing appropriate indices by 1.
-
-    Examples
-    --------
-    >>> num_periods = 40
-    >>> num_types = 1
-    >>> edu_starts, edu_max = [10], 20
-    >>> states, indexer = create_state_space(
-    ...     num_periods, num_types, edu_starts, edu_max
-    ... )
-    >>> states.shape
-    (317367, 6)
-    >>> indexer.shape
-    (40, 40, 40, 21, 4, 1)
-
-    """
-    data = []
-
-    shape = (num_periods, num_periods, num_periods, edu_max + 1, 4, num_types)
-    indexer = np.full(shape, -1, dtype=np.int32)
-
-    # Initialize counter
-    i = 0
-
-    # Construct state space by periods
-    for period in range(num_periods):
-
-        # Loop over all unobserved types
-        for type_ in range(num_types):
-
-            # Loop overall all initial levels of schooling
-            for edu_start in edu_starts:
-
-                # For occupations and education it is necessary to loop over period
-                # + 1 as zero has to be included if it is never this choice and period
-                # + 1 if it is always the same choice.
-
-                # Furthermore, the time constraint of agents is always fulfilled as
-                # previous choices limit the space for subsequent choices.
-
-                # Loop over all admissible work experiences for Occupation A
-                for exp_a in range(period + 1):
-
-                    # Loop over all admissible work experience for Occupation B
-                    for exp_b in range(period + 1 - exp_a):
-
-                        # Loop over all admissible additional education levels
-                        for edu_add in range(
-                            min(period + 1 - exp_a - exp_b, edu_max + 1 - edu_start)
-                        ):
-
-                            # Loop over all admissible values for the lagged activity:
-                            # (1) Occupation A, (2) Occupation B, (3) Education, and (4)
-                            # Home.
-                            for choice_lagged in range(4):
-
-                                if period > 0:
-
-                                    # (0, 1) Whenever an agent has only worked in
-                                    # Occupation A, then the lagged choice cannot be
-                                    # anything other than one.
-                                    if choice_lagged != 0 and exp_a == period:
-                                        continue
-
-                                    # (0, 2) Whenever an agent has only worked in
-                                    # Occupation B, then the lagged choice cannot be
-                                    # anything other than two
-                                    if choice_lagged != 1 and exp_b == period:
-                                        continue
-
-                                    # (0, 3) Whenever an agent has only acquired
-                                    # additional education, then the lagged choice
-                                    # cannot be anything other than three.
-                                    if choice_lagged != 2 and edu_add == period:
-                                        continue
-
-                                    # (0, 4) Whenever an agent has not acquired any
-                                    # additional education and we are not in the first
-                                    # period, then lagged activity cannot take a value
-                                    # of three.
-                                    if choice_lagged == 2 and edu_add == 0:
-                                        continue
-
-                                    # (0, 5) Whenever an agent has always chosen
-                                    # Occupation A, Occupation B or education, then
-                                    # lagged activity cannot take a value of four.
-                                    if (
-                                        choice_lagged == 3
-                                        and exp_a + exp_b + edu_add == period
-                                    ):
-                                        continue
-
-                                # (2, 1) An individual that has never worked in
-                                # Occupation A cannot have that lagged activity.
-                                if choice_lagged == 0 and exp_a == 0:
-                                    continue
-
-                                # (3, 1) An individual that has never worked in
-                                # Occupation B cannot have a that lagged activity.
-                                if choice_lagged == 1 and exp_b == 0:
-                                    continue
-
-                                # (1, 1) In the first period individual either were in
-                                # school the previous period as well or at home. They
-                                # cannot have any work experience.
-                                if period == 0:
-                                    if choice_lagged in [0, 1]:
-                                        continue
-
-                                # Continue if state still exist. This condition is only
-                                # triggered by multiple initial levels of education.
-                                if (
-                                    indexer[
-                                        period,
-                                        exp_a,
-                                        exp_b,
-                                        edu_start + edu_add,
-                                        choice_lagged,
-                                        type_,
-                                    ]
-                                    != -1
-                                ):
-                                    continue
-
-                                # Collect mapping of state space to array index.
-                                indexer[
-                                    period,
-                                    exp_a,
-                                    exp_b,
-                                    edu_start + edu_add,
-                                    choice_lagged,
-                                    type_,
-                                ] = i
-
-                                i += 1
-
-                                # Store information in a dictionary and append to data.
-                                row = (
-                                    period,
-                                    exp_a,
-                                    exp_b,
-                                    edu_start + edu_add,
-                                    choice_lagged,
-                                    type_,
-                                )
-                                data.append(row)
-
-    states = np.array(data)
-
-    return states, indexer
-
-
-def create_reward_components(types, covariates, optim_paras):
-    """Calculate systematic rewards for each state.
-
-    Wages are only available for some choices, i.e. n_nonpec >= n_wages. We extend the
-    array of wages with ones for the difference in dimensions, n_nonpec - n_wages. Ones
-    are necessary as it facilitates the aggregation of reward components in
-    :func:`calculate_emax_value_functions` and related functions.
-
-    Parameters
-    ----------
-    types : np.ndarray
-        Array with shape (n_states,) containing type information.
-    covariates : dict
-        Dictionary with covariate arrays for wage and nonpec rewards
-    optim_paras : dict
-        Contains parameters affected by the optimization.
-
-    """
-    wage_labels = ["wage_a", "wage_b"]
-    log_wages = np.column_stack(
-        [np.dot(covariates[w], optim_paras[w]) for w in wage_labels]
-    )
-
-    nonpec_labels = ["nonpec_a", "nonpec_b", "nonpec_edu", "nonpec_home"]
-    nonpec = np.column_stack(
-        [np.dot(covariates[n], optim_paras[n]) for n in nonpec_labels]
-    )
-
-    type_deviations = optim_paras["type_shifts"][types]
-
-    log_wages += type_deviations[:, :2]
-    nonpec[:, 2:] += type_deviations[:, 2:]
-
-    wages = np.clip(np.exp(log_wages), 0.0, HUGE_FLOAT)
-
-    # Extend wages to dimension of non-pecuniary rewards.
-    additional_dim = nonpec.shape[1] - log_wages.shape[1]
-    wages = np.column_stack((wages, np.ones((wages.shape[0], additional_dim))))
-
-    return wages, nonpec
 
 
 def solve_with_backward_induction(state_space, interpolation_points, optim_paras):
@@ -568,193 +331,6 @@ def check_prediction_model(predictions_diff, beta):
     assert np.all(np.isfinite(beta))
 
 
-def create_choice_covariates(covariates_df, states_df, params_spec):
-    """Create the covariates for each  choice.
-
-    Parameters
-    ----------
-    covariates_df : pd.DataFrame
-        DataFrame with the basic covariates.
-    states_df : pd.DataFrame
-        DataFrame with the state information.
-    params_spec : pd.DataFrame
-        The parameter specification.
-
-    Returns
-    -------
-    wage_covariates: list
-        List of length nchoices with covariate arrays for systematic wages.
-    nonpec_covariates: list
-        List of length nchoices with covariate arrays for nonpecuniary rewards.
-
-    """
-    all_data = pd.concat([covariates_df, states_df], axis="columns", sort=False)
-
-    covariates = {}
-
-    for choice in ["a", "b", "edu", "home"]:
-        if f"wage_{choice}" in params_spec.index:
-            wage_columns = params_spec.loc[f"wage_{choice}"].index
-            covariates[f"wage_{choice}"] = all_data[wage_columns].to_numpy()
-
-        nonpec_columns = params_spec.loc[f"nonpec_{choice}"].index
-        covariates[f"nonpec_{choice}"] = all_data[nonpec_columns].to_numpy()
-
-    for key, val in covariates.items():
-        covariates[key] = np.ascontiguousarray(val)
-
-    return covariates
-
-
-class StateSpace:
-    """Class containing all objects related to the state space of a discrete choice
-    dynamic programming model.
-
-    Parameters
-    ----------
-    attr : dict
-        Dictionary containing model attributes.
-    optim_paras : dict
-        Dictionary containing parameters affected by optimization.
-
-    Attributes
-    ----------
-    states : np.ndarray
-        Array with shape (num_states, 6) containing period, exp_a, exp_b, edu,
-        choice_lagged and type information.
-    indexer : np.ndarray
-        Array with shape (num_periods, num_periods, num_periods, edu_max, n_choices,
-        num_types).
-    covariates : np.ndarray
-        Array with shape (num_states, 16) containing covariates of each state necessary
-        to calculate rewards.
-    wages : np.ndarray
-        Array with shape (n_states_in_period, n_choices) which contains zeros in places
-        for choices without wages.
-    nonpec : np.ndarray
-        Array with shape (n_states_in_period, n_choices).
-    emaxs : np.ndarray
-        Array with shape (num_states, 5) containing containing the emax of each choice
-        (OCCUPATION A, OCCUPATION B, SCHOOL, HOME) of the subsequent period and the
-        simulated or interpolated maximum of the current period.
-    num_periods : int
-        Number of periods.
-    num_types : int
-        Number of types.
-    states_columns : list
-        List of column names in ``self.states``.
-    base_covariates_columns : list
-        List of column names in ``self.covariates``.
-    wages_columns : list
-        List of column names in ``self.wages``
-    nonpec_columns : list
-        List of column names in ``self.nonpec``
-    emaxs_columns : list
-        List of column names in ``self.continuation_values``.
-
-    """
-
-    states_columns = ["period", "exp_a", "exp_b", "edu", "choice_lagged", "type"]
-
-    base_covariates_columns = list(BASE_COVARIATES.keys())
-
-    wages_columns = ["wage_a", "wage_b"]
-    nonpec_columns = ["nonpec_a", "nonpec_b", "nonpec_edu", "nonpec_home"]
-    emaxs_columns = ["emax_a", "emax_b", "emax_edu", "emax_home", "emax"]
-
-    def __init__(self, params, options):
-        params, optim_paras = process_params(params)
-        options = process_options(options)
-
-        # Add some arguments to the state space.
-        self.edu_max = options["education_max"]
-        self.num_periods = options["num_periods"]
-        self.num_types = optim_paras["num_types"]
-        self.seed = options["solution_seed"]
-
-        self.base_draws_sol = create_base_draws(
-            (self.num_periods, options["solution_draws"], 4), self.seed
-        )
-
-        self.states, self.indexer = create_state_space(
-            self.num_periods, self.num_types, options["education_start"], self.edu_max
-        )
-        states_df = pd.DataFrame(data=self.states, columns=self.states_columns)
-
-        base_covariates_df = create_base_covariates(states_df, options["covariates"])
-
-        self.covariates = create_choice_covariates(
-            base_covariates_df, states_df, params
-        )
-
-        self.wages, self.nonpec = create_reward_components(
-            self.states[:, 5], self.covariates, optim_paras
-        )
-
-        self.is_inadmissible = create_inadmissible_indicator(states_df, options)
-
-        self._create_slices_by_periods(self.num_periods)
-
-    @property
-    def states_per_period(self):
-        """Get a list of states per period starting from the first period."""
-        return np.array([len(range(i.start, i.stop)) for i in self.slices_by_periods])
-
-    @property
-    def num_states(self):
-        """Get the total number states in the state space."""
-        return self.states.shape[0]
-
-    def update_systematic_rewards(self, optim_paras):
-        self.wages, self.nonpec = create_reward_components(
-            self.states[:, 5], self.covariates, optim_paras
-        )
-
-    def get_attribute_from_period(self, attr, period):
-        """Get an attribute of the state space sliced to a given period.
-
-        Parameters
-        ----------
-        attr : str
-            String of attribute name, e.g. ``"states"``.
-        period : int
-            Attribute is retrieved from this period.
-
-        """
-        if attr == "covariates":
-            raise AttributeError("Attribute covariates cannot be retrieved by periods.")
-        else:
-            pass
-
-        try:
-            attribute = getattr(self, attr)
-        except AttributeError as e:
-            raise AttributeError(f"StateSpace has no attribute {attr}.").with_traceback(
-                e.__traceback__
-            )
-
-        try:
-            indices = self.slices_by_periods[period]
-        except IndexError as e:
-            raise IndexError(f"StateSpace has no period {period}.").with_traceback(
-                e.__traceback__
-            )
-
-        return attribute[indices]
-
-    def _create_slices_by_periods(self, num_periods):
-        """Create slices to index all attributes in a given period.
-
-        It is important that the returned objects are not fancy indices. Fancy indexing
-        results in copies of array which decrease performance and raise memory usage.
-
-        """
-        self.slices_by_periods = []
-        for i in range(num_periods):
-            idx_start, idx_end = np.where(self.states[:, 0] == i)[0][[0, -1]]
-            self.slices_by_periods.append(slice(idx_start, idx_end + 1))
-
-
 @guvectorize(
     [
         "f4[:], f4[:], f4[:], f4[:, :], f4, b1[:], f4[:]",
@@ -924,52 +500,6 @@ def calculate_value_functions(
             )
 
             value_functions[j, i] = value_function
-
-
-def create_base_covariates(states, covariates_spec):
-    """Create set of covariates for each state.
-
-    Parameters
-    ----------
-    states : np.ndarray
-        Array with shape (num_states, 6) containing period, exp_a, exp_b, edu,
-        choice_lagged and type of each state.
-    covariates_spec : dict
-        Dictionary where keys represent covariates and values are strings which can be
-        passed to pd.eval.
-
-    Returns
-    -------
-    covariates : np.ndarray
-        Array with shape (num_states, 16) containing covariates of each state.
-
-    """
-    covariates = states.copy()
-
-    for covariate, definition in covariates_spec.items():
-        covariates[covariate] = covariates.eval(definition)
-
-    covariates = covariates.drop(columns=states.columns).astype(float)
-
-    return covariates
-
-
-def create_inadmissible_indicator(states, options):
-    df = states.copy()
-
-    # Add maximum of experiences.
-    for key in options.keys():
-        if "max" in key:
-            df[key] = options[key]
-        else:
-            pass
-
-    for column, definition in options["inadmissible_states"].items():
-        df[column] = df.eval(definition)
-
-    is_inadmissible = df[options["inadmissible_states"].keys()].to_numpy()
-
-    return is_inadmissible
 
 
 def ols(y, x):
