--- respy/tests/test_likelihood.py
@@ -1,1 +1,0 @@
-"""Test the likelihood routine."""
